round and around

+++ fft_convolve_cuda.cu
@@ -56,11 +56,12 @@ cudaProdScaleKernel(const cufftComplex *raw_data, const cufftComplex *impulse_v,
     while(idx < padded_length){
       float a = raw_data[idx].x;
       float b = raw_data[idx].y;
+
       float c = impulse_v[idx].x;
       float d = impulse_v[idx].y;
 
-      out_data[idx].x = (a * c - b * d) / padded_length;
-      out_data[idx].y = (a * d - b * c) / padded_length;
+      out_data[idx].x = (a * c - b * d) / (float) padded_length;
+      out_data[idx].y = (a * d - b * c) / (float) padded_length;
 
       idx += blockDim.x * gridDim.x;
     }
